day1_task2/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
groq==0.4.1
streamlit==1.28.1
python-dotenv==1.0.0
pydantic==1.10.13